add toastify

+++ src/controllers/product/main.js
@@ -1,8 +1,7 @@
-import electron from "electron";
+import { ipcMain } from "electron";
 import { database } from "../../db.js";
 import { ModelProduct } from "../../models/ModelProduct.js";
-
-const { ipcMain } = electron;
+import { ModelProductOrder } from "../../models/ModelProductOrder.js";
 
 ipcMain.on("createProduct", async (event, arg) => {
   try {
@@ -20,7 +19,7 @@ ipcMain.on("createProduct", async (event, arg) => {
     event.reply("createProduct-reply", result.dataValues);
   } catch (error) {
     console.log(error);
-    event.reply("createProduct-reply", error.message);
+    event.reply("createProduct-reply", { error: error.message });
   }
 });
 
@@ -34,7 +33,7 @@ ipcMain.on("getAllProduct", async (event, _) => {
     event.reply("getAllProductl-reply", result);
   } catch (error) {
     console.log(error);
-    event.reply("getAllProduct-reply", error.message);
+    event.reply("getAllProduct-reply", { error: error.message });
   }
 });
 
@@ -50,7 +49,7 @@ ipcMain.on("getAllProduct", async (event, _) => {
     event.reply("getAllProduct-reply", result);
   } catch (error) {
     console.log(error);
-    event.reply("getAllProduct-reply", error.message);
+    event.reply("getAllProduct-reply", { error: error.message });
   }
 });
 
@@ -63,19 +62,30 @@ ipcMain.on("editProduct", async (event, arg) => {
     event.reply("editProduct-reply", product);
   } catch (error) {
     console.log(error);
-    event.reply("editProduct-reply", error.message);
+    event.reply("editProduct-reply", { error: error.message });
   }
 });
 
 ipcMain.on("deleteProduct", async (event, arg) => {
   try {
     await database.sync();
+    const existsOrder = await ModelProductOrder.findOne({
+      where: { productId: arg.id },
+      raw: true,
+    });
+
+    if (existsOrder) {
+      throw new Error(
+        "Não é possível deletar produto, há registros dele em pedidos!"
+      );
+    }
+
     const product = await ModelProduct.destroy({ where: { id: arg.id } });
-    if (!product) throw new Error("Não foi possível fazer o registro!");
+    if (!product) throw new Error("Não foi possível deletar esse produto!");
 
     event.reply("deleteProduct-reply", { id: arg.id });
   } catch (error) {
     console.log(error);
-    event.reply("deleteProduct-reply", error.message);
+    event.reply("deleteProduct-reply", { error: error.message });
   }
 });
